Fix pathing to update dispatch file in correct dir
As it stands the current script would create/update a .dispatch file in `Deployment\Resources\Dispatch` rather than `Deployment\Resources\Dispatch\locale-folder` which means that when the dispatch refresh command is run (and looks for a dispatch file inside the locale-folder) it won't be the file that was just updated. This change update the behaviour so that we are both writing to, and reading from the locale-folder.